From previous commit, instead move list_email_recipients over to the email provider.
